Shifted profile management to Account.tsx
